requirements.txt
asgiref==3.5.2
Django==4.1.4
django-crispy-forms==1.14.0
django-ipware==4.0.2
itermate==1.0.2
Pillow==9.3.0
somewhere==0.0.1
tzdata==2022.7
whitenoise==6.2.0
